Added isAdmin fn into the auth middleware file to validate that a user is of a role that we specified guest or admin.
In this case we pass Admin into the params of the middleware, then we added the middleware for the movies route, only admin user
can create, update and delete a movie.

Also we added the isAdmin middleware into the users route, so only admin user can fetch all the data
on the users table. In order to give a user the role of admin, we are going to allow
that other admin user can give admin role to a "default" or "guest" user by using the endpoint /admin and
the controller updateUserRole.