Related our GET (Movies and actors) requests, either by ID or fetch all resources to properly show our relations.
We mapped through the data and convert the base URL that firebase
provide us to a public domain URL to further consume it. Also we made a
middleware in the auth file, to protect the PATCH Request on the
accounts of users. Only the current user can edit their data and
no one else. At the end we add also run prettier in all of our files.