Old project commit, working in dev env, but never uploaded to remote.
Since the teacher requirements did not involved uploading it to a
hosting service ex. Heroku. I only demo the project working in dev
environment and pass the assignment.

This commit includes all the changes in the codebase related to routes
guard middlewares, some changes in the schema and email setup with
sendgrid and nodemailer.